leetcode/101-200/0107.Binary-Tree-Level-Order-Traversal-II/Solution_test.go
package Solution

import (
	"reflect"
	"testing"
)

func TestSolution(t *testing.T) {
	//	测试用例
	cases := []struct {
		name   string
		inputs *TreeNode
		expect [][]int
	}{
		{"1 test 1", &TreeNode{3, &TreeNode{Val: 9}, &TreeNode{20, &TreeNode{Val: 15}, &TreeNode{Val: 7}}}, [][]int{{15, 7}, {9, 20}, {3}}},
		{"2 test 2", &TreeNode{Val: 1}, [][]int{{1}}},
		{"3 test 3", &TreeNode{}, [][]int{{0}}},
	}

	//	开始测试
	for _, c := range cases {
		t.Run(c.name, func(t *testing.T) {
			ret := Solution(c.inputs)
			if !reflect.DeepEqual(ret, c.expect) {
				t.Fatalf("expected: %v, but got: %v, with inputs: %v",
					c.expect, ret, c.inputs)
			}
		})
	}
}
